fix runtime api, reorg proposals
