+++ pallets/proposals/src/impls/mod.rs
@@ -1,0 +1,6 @@
+
+pub mod pallet_impls;
+pub mod immutable_votes;
+
+pub use pallet_impls::*;
+pub use immutable_votes::*;
